requirementes.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2
django-filter==22.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
djangorestframework-simplejwt

